Bot_final/requirements.txt
maxgram
psycopg2-binary
pymorphy3
pymorphy3-dicts-ru
requests
beautifulsoup4
urllib3
python-dateutil







